i broke the training, restarted and changed how it's done slightly

+++ StreetViewImageRetrieve/Classes/PanoPosition.cs
@@ -2,11 +2,11 @@
 {
     public struct PanoPosition
     {
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
 
-        public PanoPosition(float latitude, float longitude)
+        public PanoPosition(double latitude, double longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
